updated task_respository functions and started working on the Task class in the task_model file
