+++ api/requirements.txt
@@ -5,4 +5,3 @@ Jinja2==2.11.2
 MarkupSafe==1.1.1
 Werkzeug==1.0.1
 pymongo==3.11.0
-Flask-PyMongo==2.3.0
