+++ api/src/task/task_model.py
@@ -1,4 +1,10 @@
 class Task:
 
-    def __init__(self):
-        pass
+    # wszystkie atrybuty jakie musi mieć task
+    def __init__(self, name, description, user_id, goal_id, state):
+        self.name = name
+        self.description = description
+        self.user_id = user_id
+        self.goal_id = goal_id
+        self.state = state
+
