+++ api/src/task/task_service.py
@@ -2,4 +2,6 @@ from .task_model import Task
 from .task_repository import TaskRepository
 
 class TaskService:
-    pass
+    
+    def __init__(self, taskRepository):
+        pass
